More typing + refactoring of __init__ to client + client_factory
